correção de erro na api - API jwt

+++ jwt/src/main/java/com/security/jwt/service/UsuarioService.java
@@ -25,7 +25,7 @@ public class UsuarioService {
         return usuarioRepository.save(user);
     }
 
-    public Optional<Usuario> findyById(Long id) {
+    public Optional<Usuario> findyById(Integer id) {
         return usuarioRepository.findById(id);
     }
 }
